// src/components/fire/types/FireForm.types.ts
export interface UseInputFormReturn {
  inputs: FireInputs
  updateBase: <K extends keyof FireInputs["baseProfile"]>(key: K, value: FireInputs["baseProfile"][K]) => void
  updateExpense: <K extends keyof FireInputs["expenseProfile"]>(key: K, value: FireInputs["expenseProfile"][K]) => void
  updateInvestment: <K extends keyof FireInputs["investmentProfile"]>(key: K, value: FireInputs["investmentProfile"][K]) => void
  updateChild: <K extends keyof ChildProfile>(index: number, key: K, value: ChildProfile[K]) => void
  addChild: () => void
  removeChild: (index: number) => void
  updateEducation: (stage: "school" | "graduation" | "postGraduation", key: string, value: number) => void
  updateMarriage: <K extends keyof FireInputs["marriageParameters"]>(key: K, value: FireInputs["marriageParameters"][K]) => void
  updateWhitegood: <K extends keyof WhitegoodsItem>(index: number, key: K, value: WhitegoodsItem[K]) => void
  addWhitegood: () => void
  removeWhitegood: (index: number) => void
  updateTravel: <K extends keyof FireInputs["travelParameters"]>(key: K, value: FireInputs["travelParameters"][K]) => void
  updateHealthcare: <K extends keyof FireInputs["healthcareParameters"]>(key: K, value: FireInputs["healthcareParameters"][K]) => void
  updateFireAssumptions: <K extends keyof FireInputs["fireAssumptions"]>(key: K, value: FireInputs["fireAssumptions"][K]) => void
  updateGoalInvestment: <K extends keyof GoalInvestmentProfile>(key: K, value: GoalInvestmentProfile[K]) => void
  updateGlidepath: (index: number, key: keyof GlidepathCheckpoint, value: number) => void
  addGlidepointCheckpoint: () => void
  removeGlidepointCheckpoint: (index: number) => void
  fileInputRef: React.RefObject<HTMLInputElement | null>
  handleExport: () => void
  handleImport: (e: React.ChangeEvent<HTMLInputElement>) => void
  handleCalculate: () => void
}